Rename HttpExecution[Context] to ClassLoaderExecution[Context]
