+++ core/play/src/main/java/play/libs/concurrent/ClassLoaderExecution.java
@@ -5,15 +5,15 @@
 package play.libs.concurrent;
 
 import java.util.concurrent.Executor;
-import play.core.j.HttpExecutionContext;
+import play.core.j.ClassLoaderExecutionContext;
 import scala.concurrent.ExecutionContext;
 import scala.concurrent.ExecutionContextExecutor;
 
 /**
  * ExecutionContexts that preserve the current thread's context ClassLoader by passing it through
- * {@link play.libs.concurrent.HttpExecutionContext}.
+ * {@link play.libs.concurrent.ClassLoaderExecutionContext}.
  */
-public class HttpExecution {
+public class ClassLoaderExecution {
 
   /**
    * An ExecutionContext that executes work on the given ExecutionContext. The current thread's
@@ -21,10 +21,11 @@ public class HttpExecution {
    * tasks.
    *
    * @param delegate the delegate execution context.
-   * @return the execution context wrapped in an {@link play.libs.concurrent.HttpExecutionContext}.
+   * @return the execution context wrapped in an {@link
+   *     play.libs.concurrent.ClassLoaderExecutionContext}.
    */
   public static ExecutionContextExecutor fromThread(ExecutionContext delegate) {
-    return HttpExecutionContext.fromThread(delegate);
+    return ClassLoaderExecutionContext.fromThread(delegate);
   }
 
   /**
@@ -33,10 +34,11 @@ public class HttpExecution {
    * tasks.
    *
    * @param delegate the delegate execution context.
-   * @return the execution context wrapped in an {@link play.libs.concurrent.HttpExecutionContext}.
+   * @return the execution context wrapped in an {@link
+   *     play.libs.concurrent.ClassLoaderExecutionContext}.
    */
   public static ExecutionContextExecutor fromThread(ExecutionContextExecutor delegate) {
-    return HttpExecutionContext.fromThread(delegate);
+    return ClassLoaderExecutionContext.fromThread(delegate);
   }
 
   /**
@@ -45,9 +47,10 @@ public class HttpExecution {
    * tasks.
    *
    * @param delegate the delegate execution context.
-   * @return the execution context wrapped in an {@link play.libs.concurrent.HttpExecutionContext}.
+   * @return the execution context wrapped in an {@link
+   *     play.libs.concurrent.ClassLoaderExecutionContext}.
    */
   public static ExecutionContextExecutor fromThread(Executor delegate) {
-    return HttpExecutionContext.fromThread(delegate);
+    return ClassLoaderExecutionContext.fromThread(delegate);
   }
 }
