+++ core/play/src/main/java/play/libs/concurrent/CustomExecutionContext.java
@@ -62,6 +62,6 @@ public abstract class CustomExecutionContext implements ExecutionContextExecutor
    * @return This executor that will execute its tasks with the current ClassLoader.
    */
   public Executor current() {
-    return HttpExecution.fromThread(this);
+    return ClassLoaderExecution.fromThread(this);
   }
 }
