+++ documentation/manual/working/commonGuide/configuration/ThreadPools.md
@@ -43,9 +43,9 @@ In most situations, the appropriate execution context to use will be the **Play 
 
 @[global-thread-pool](code/ThreadPools.scala)
 
-or using [`CompletionStage`](https://docs.oracle.com/javase/8/docs/api/java/util/concurrent/CompletionStage.html) with an [`HttpExecutionContext`](api/java/play/libs/concurrent/HttpExecutionContext.html) in Java code:
+or using [`CompletionStage`](https://docs.oracle.com/javase/8/docs/api/java/util/concurrent/CompletionStage.html) with an [`ClassLoaderExecutionContext`](api/java/play/libs/concurrent/ClassLoaderExecutionContext.html) in Java code:
 
-@[http-execution-context](code/detailedtopics/httpec/MyController.java)
+@[cl-execution-context](code/detailedtopics/clec/MyController.java)
 
 This execution context connects directly to the Application's `ActorSystem` and uses Akka's [default dispatcher][akka-default-dispatcher].
 
@@ -103,13 +103,13 @@ In some cases you may not be able to explicitly use the application classloader.
 
 ### Switching threads
 
-The problem with class loaders however is that as soon as control switches to another thread, you lose access to the original class loader. So if you were to map a `CompletionStage` using `thenApplyAsync`, or using `thenApply` at a point in time after the `Future` associated with that `CompletionStage` had completed, and you then try to access the original class loader, it probably won't work .  To address this, Play provides an [`HttpExecutionContext`](api/java/play/libs/concurrent/HttpExecutionContext.html).  This allows you to capture the current class loader in an `Executor`, which you can then pass to the `CompletionStage` `*Async` methods such as `thenApplyAsync()`, and when the executor executes your callback, it will ensure the class loader remains in scope.
+The problem with class loaders however is that as soon as control switches to another thread, you lose access to the original class loader. So if you were to map a `CompletionStage` using `thenApplyAsync`, or using `thenApply` at a point in time after the `Future` associated with that `CompletionStage` had completed, and you then try to access the original class loader, it probably won't work .  To address this, Play provides an [`ClassLoaderExecutionContext`](api/java/play/libs/concurrent/ClassLoaderExecutionContext.html).  This allows you to capture the current class loader in an `Executor`, which you can then pass to the `CompletionStage` `*Async` methods such as `thenApplyAsync()`, and when the executor executes your callback, it will ensure the class loader remains in scope.
 
-To use the `HttpExecutionContext`, inject it into your component, and then pass the current execution context anytime a `CompletionStage` is interacted with.  For example:
+To use the `ClassLoaderExecutionContext`, inject it into your component, and then pass the current execution context anytime a `CompletionStage` is interacted with.  For example:
 
-@[http-execution-context](code/detailedtopics/httpec/MyController.java)
+@[cl-execution-context](code/detailedtopics/clec/MyController.java)
 
-If you have a custom executor, you can wrap it in an `HttpExecutionContext` simply by passing it to the `HttpExecutionContext`s constructor.
+If you have a custom executor, you can wrap it in an `ClassLoaderExecutionContext` simply by passing it to the `ClassLoaderExecutionContext`s constructor.
 
 ## Best practices
